refactor: implement ListQuestionService and update ListQuestionController
